Rename Metropolis Module
