+++ src/minigraphs/__init__.py
@@ -4,7 +4,7 @@ Minigraphs
 
 Python package to miniaturize large networks using the Metropolis-Hastings Algorithm
 """
-import minigraphs.Metropolis
+from minigraphs.miniaturize import *
 import minigraphs.io 
 import minigraphs.visualization
